Update: Migrations

+++ egressos-backend/database/migrations/2024_09_18_123400_create_platforms_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::create('platforms', function (Blueprint $table) {
-            $table->increments('id'); // Alterado para integer() em vez de id()
+            $table->bigIncrements('id'); // Alterado para biginteger() em vez de id()
             $table->string('name');
             $table->timestamps();
         });
